Small cleanups in actor_ref module
Removes the ActorShutdown error, as this is no longer used. And moves
the SendError into the src/actor_ref/mod file.
